refactoring

+++ API/Web.API/Controllers/UserController.cs
@@ -23,7 +23,7 @@ namespace Web.API.Controllers
         {
             try
             {
-                await _userService.RegisterUserAsync(model);
+                await _userService.AddUserAsync(model);
             }
             catch (InvalidOperationException)
             {
@@ -52,7 +52,7 @@ namespace Web.API.Controllers
             };
 
             Response.Cookies.Append("AppCookie", token, cookieOptions);
-            return Ok();
+            return NoContent();
         }
 
         [Authorize]
@@ -64,61 +64,72 @@ namespace Web.API.Controllers
                 Response.Cookies.Delete("AppCookie");
             }
 
-            return Ok();
+            return NoContent();
         }
 
         [Authorize]
         [HttpGet("searchByEmail")]
         public async Task<IActionResult> SearchByEmail(string email)
         {
-            var currentUserId = Guid.Parse(User.Claims.FirstOrDefault(x => x.Type == "userId").Value);
-            var users = await _userService.SearchByEmailAsync(email, currentUserId);
-            return Ok(users);
+            var userIdClaim = User?.Claims.FirstOrDefault(x => x.Type == "userId")?.Value;
+            if (Guid.TryParse(userIdClaim, out Guid currentUserId))
+            {
+                var users = await _userService.SearchByEmailAsync(email, currentUserId);
+                return Ok(users);
+            }
+            return Unauthorized();
         }
 
         [Authorize]
         [HttpGet("id")]
         public IActionResult GetUserId()
         {
-            var currentUserId = Guid.Parse(User.Claims.FirstOrDefault(x => x.Type == "userId").Value);
-            return Ok(currentUserId);
+            var userIdClaim = User?.Claims.FirstOrDefault(x => x.Type == "userId")?.Value;
+            if (Guid.TryParse(userIdClaim, out Guid currentUserId))
+            {
+                return Ok(currentUserId);
+            }
+            return Unauthorized();
         }
 
         [Authorize]
         [HttpPut("profileImage")]
         public async Task<IActionResult> UpdateProfileImage(IFormFile file)
         {
-            var currentUserId = Guid.Parse(User.Claims.FirstOrDefault(x => x.Type == "userId").Value);
-
-            var validExtensions = new[] { ".jpg", ".jpeg", ".png", ".gif", ".bmp" };
-            var extension = Path.GetExtension(file.FileName).ToLowerInvariant();
-
-            var validContentTypes = new[] { "image/jpg", "image/jpeg", "image/png", "image/gif", "image/bmp" };
-            var contentType = file.ContentType.ToLowerInvariant();
-
-            if (!validExtensions.Contains(extension) || !validContentTypes.Contains(contentType))
+            var userIdClaim = User?.Claims.FirstOrDefault(x => x.Type == "userId")?.Value;
+            if (Guid.TryParse(userIdClaim, out Guid currentUserId))
             {
-                return StatusCode(409);
+                var validExtensions = new[] { ".jpg", ".jpeg", ".png", ".gif", ".bmp" };
+                var extension = Path.GetExtension(file.FileName).ToLowerInvariant();
+
+                var validContentTypes = new[] { "image/jpg", "image/jpeg", "image/png", "image/gif", "image/bmp" };
+                var contentType = file.ContentType.ToLowerInvariant();
+
+                if (!validExtensions.Contains(extension) || !validContentTypes.Contains(contentType))
+                {
+                    return StatusCode(409);
+                }
+
+                var oldFiles = Directory.GetFiles(Path.Combine(_folderPath, "profile-images"), $"{currentUserId}.*");
+
+                foreach (var oldFile in oldFiles)
+                {
+                    System.IO.File.Delete(oldFile);
+                }
+
+                var path = $"profile-images/{currentUserId}{extension}";
+                var filePath = Path.Combine(_folderPath, path);
+
+                using (var stream = new FileStream(filePath, FileMode.Create, FileAccess.Write, FileShare.None))
+                {
+                    await file.CopyToAsync(stream);
+                }
+
+                await _userService.UpdateProfileAsync(path, currentUserId);
+
+                return NoContent();
             }
-
-            var oldFiles = Directory.GetFiles(Path.Combine(_folderPath, "profile-images"), $"{currentUserId}.*");
-
-            foreach (var oldFile in oldFiles)
-            {
-                System.IO.File.Delete(oldFile);
-            }
-
-            var path = $"profile-images/{currentUserId}{extension}";
-            var filePath = Path.Combine(_folderPath, path);
-
-            using (var stream = new FileStream(filePath, FileMode.Create, FileAccess.Write, FileShare.None))
-            {
-                await file.CopyToAsync(stream);
-            }
-
-            await _userService.UpdateProfileAsync(path, currentUserId);
-
-            return NoContent();
+            return Unauthorized();
         }
 
         [Authorize]
@@ -137,9 +148,13 @@ namespace Web.API.Controllers
         [HttpGet]
         public async Task<IActionResult> GetCurrentUser()
         {
-            var currentUserId = Guid.Parse(User.Claims.FirstOrDefault(x => x.Type == "userId").Value);
-            var user = await _userService.GetUserAsync(currentUserId);
-            return Ok(user);
+            var userIdClaim = User?.Claims.FirstOrDefault(x => x.Type == "userId")?.Value;
+            if (Guid.TryParse(userIdClaim, out Guid currentUserId))
+            {
+                var user = await _userService.GetUserAsync(currentUserId);
+                return Ok(user);
+            }
+            return Unauthorized();
         }
     }
 }
